+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "DONT COMMIT JUST SAVE" extension will be documented 
 
 -   `post-rewrite` and `post-checkout` hooks
 -   Startup marker warning
+-   Custom count option when consecutive "DONT COMMIT JUST SAVE" commits are detected in soft reset
 
 ### Changed